requirements.txt
beautifulsoup4>=4.9.3
bs4>=0.0.1
greenlet>=1.1.0
importlib-metadata>=4.5.0
lxml>=4.6.3
numpy>=1.20.3
pandas>=1.2.4
psycopg2-binary>=2.9.1
python-dateutil>=2.8.1
python-dotenv
pytz>=2021.1
requests
six
SQLAlchemy>=2.0
wget>=3.2
zipp>=3.4.1
dask
pyarrow
